extract out leaderboard aggregation
